Move main::authenticate to the top level of the library
Also clean up a few dependencies that the binary had that were
transitive dependencies only
