+++ lib/Cargo.toml
@@ -11,6 +11,7 @@ hex = "0.4.3"
 hmac = "0.12.1"
 librespot-core = "0.3.1"
 librespot-protocol = "0.3.1"
+log = "0.4"
 minreq = { version = "2.6", default-features = false, features = ['urlencoding'] }
 pbkdf2 = "0.11"
 rand = "0.8.5"
